feat: prevent sql injection
