--- src/app/api/v1/hello/route.ts
@@ -1,5 +1,0 @@
-import { NextResponse } from 'next/server'
-
-export async function GET() {
-  return NextResponse.json({ message: 'Hello, world!' })
-}
